Let the indices be indices: size_t where they only index
The casts the container change left behind were the wrong shape: 67 to
int32 from size(), 72 the other way, the same index converted back and
forth. Two measurements decided the rule rather than taste.

Descending loops that an unsigned index would break: two, and both are
"for (i = 15; i > 7; i--)" over the Moore neighbourhood directions in the
contour tracer - not container indices at all. Sentinels of -1: one, which
I had introduced myself the day before while translating indexOf.

So indices become std::size_t, because that removes the casts on BOTH
sides: a signed index with an ssize() helper would clean the loop bound
and leave every at(static_cast<size_t>(i)) behind.

Counts stay std::int32_t, and that is not an inconsistency. They are
validated user input - the grid dialog refuses fewer than two points - and
they do arithmetic where an unsigned underflow would silently become a
huge bound: BoundaryUnion1D::bucketIndex clamps with "res > phaseCount -
1", which for phaseCount = 0 unsigned is no clamp at all. The boundary
between the two kinds leaves one cast, and that one says something.

The frequency index that crosses the detector's API is a size_t now, 28
loops lost their cast bound, and 41 redundant casts are gone: 63 down to
22, the remainder being over expressions rather than a bare name.

Verified rather than assumed, because zero warnings was not evidence:
-Wall does not include -Wsign-conversion, so silently converting a signed
int32 would not have said anything. An explicit pass with it over the 71
core and persistence units gives 336 before this change and 310 after - 26
removed, none introduced.

Those 310 are the rest of the job. The flag has never been on in this
project, so they are pre-existing, and turning it on for real is this
item's end state: not another sweep of casts but deciding the type in the
counts too, read site by site.

+++ src/core/boundaries/boundary_union_1d.cpp
@@ -46,10 +46,10 @@ std::vector<TraceSet> BoundaryUnion1D::buildLayerBuckets(const TraceSet & chosen
 
     for (std::int32_t i = 0; i < kLayerCount; i++)
     {
-        TraceSet & row = layerBuckets[static_cast<std::size_t>(i)];
-        row.resize(static_cast<std::size_t>(totalPhase) + 1);
+        TraceSet & row = layerBuckets[i];
+        row.resize(totalPhase + 1);
 
-        for (const qftbx::NicholsPoint & point : chosenCurves.at(static_cast<std::size_t>(i)))
+        for (const qftbx::NicholsPoint & point : chosenCurves.at(i))
         {
             Trace & bucket = row.at(static_cast<std::size_t>(bucketIndex(point.phase, totalPhase)));
 
@@ -280,7 +280,7 @@ inline std::int32_t BoundaryUnion1D::bucketIndex(double x, double totalPhase, st
 
 TraceSet BoundaryUnion1D::buildUnionBuckets(const Trace & unionPoints, double totalPhase, std::int32_t pointCount)
 {
-    TraceSet unionBucketsRow (static_cast<std::size_t>(pointCount));
+    TraceSet unionBucketsRow (pointCount);
 
     //From the first point on (it used to be skipped), inserted sorted by
     //magnitude and deduplicated, like the layer buckets (and like the
