+++ src/core/specifications/specification_record.h
@@ -87,7 +87,7 @@ inline Specification toSpecification(const SpecificationRecord & d, Specificatio
 inline SpecificationSet toSpecificationSet(const SpecificationRecords & specs){
     SpecificationSet set;
     for (int i = 0; i < kSpecificationCount; ++i){
-        set.set(toSpecification(specs.at(static_cast<std::size_t>(i)),
+        set.set(toSpecification(specs.at(i),
                                 static_cast<SpecificationType>(i)));
     }
     return set;
